Change the config format to yaml
Yaml is the format that openstack-infra uses, it's easier to
read and not nearly as ugly as a json file.

It has been decided that we will replace the json formatted config
file with a yaml one. This change completely removes json config
support from turbo hipster and replaces it with yaml. As such the
old config.json has also been replaced by a yaml version.

Change-Id: Ic4c62281ee3796e5f118e441722377046e850135