Change the config format to yaml
Yaml is the format that openstack-infra uses, it's easier to
read and not nearly as ugly as a json file.

For backwards compatibility, this change doesn't remove support
for json formatted config files. Instead, if it fails to load it
using yaml it will fall back to json.

As such, I have ported the config.json over to yaml and has been
added along side the older config.json config file.

Change-Id: Ic4c62281ee3796e5f118e441722377046e850135