Fix PBR generated binary error
The PBR generated binary 'turbo-hipster' fails to run as the main
method entry point specified now requires a parameter, which PBR
doesn't handle.

This change adds a new entrypoint for PBR, still named main which
will parse the args and create the parameter needed and then called
the the setup_server method.

Change-Id: Ic11f1da0d2f9a615528bd8620ddf30d81a686498