Send a signal once the worker thread has shutdown
The work thread would shut down itself successfully if requested to
at the end of a job run. However the daemon process was waiting for
an interrupt to check the state of the thread. Add in a kill signal
once the server thread has shutdown to also shutdown the daemon.

Change-Id: I20935773518653916f8ed7d5d3734f7b9e478998