Pass all zuul arguments to the job's shell script
The idea is that Zuul passes a lot of useful information, such as the
project name, target branch etc., as ZUUL_* options within the JSON job
data. Let's pass all of these variables to the launched shell script.
This will make it possible to e.g. have a single script that can do the
"right thing" for different projects with distinct build steps, etc.

Change-Id: I6f57ed19e9eef1b987fe2493d409b3ee51af22b6