Remove the build directory during cleanup
The test results are already uploaded, so it doesn't make much sense to
keep stuff around. Because there might be some other jobs executing at
the time we get around to cleaning, the removal is rather conservative,
and it's quite possible that some empty directories are retained (for
example, the hierarchy 06/106/3/check/somejob for "somejob" for version
3 of change 106 in pipeline check). The actual content produced by the
job is removed, though.

Change-Id: Ide81c16f95dc4b41417a065640512df50bf8c8c2